+++ mockserver/mockserver-core/src/main/java/org/mockserver/validator/jsonschema/JsonSchemaExpectationValidator.java
@@ -49,6 +49,7 @@ public class JsonSchemaExpectationValidator extends JsonSchemaValidator {
             "bodyWithContentType",
             "delay",
             "connectionOptions",
+            "recoverAfter",
             "keyToMultiValue",
             "keyToValue",
             "socketAddress",